Structure of the code for the b tag reshaping SF
This commit introduces the structure in the skim code to evaluate the reshaping SF and store their normalisation.
The implementation is *not finished* due to an observed discrepancy with the previous WP SF (ongoing investigation with BTV POG).
The config files default to the application of the WP SF.
Items to do:
   * handling of systematics. Some systematics are not defined in the .csv file (e.g., lf syst for b and c jets) and currently I am taking a weight of 1 when this happens, but this does not preserve the normalisation. Should instead the nominal weight be taken? Or is the systematic a multiplicative factor on top of that?
   * histo filler must be modified to read the additional histogram with the BDT reshaping normalisation and scale the histograms accordingly
   * test all the procedure and compare the WP and reshaping SF final histograms